requirements.txt
Bio
networkx
jupyter
matplotlib>=3.4.0
mygene==3.2.2
numpy>=1.21.0
openai==1.3.9
google-generativeai>=0.3.0
pandas>=1.3.0
requests==2.31.0
scikit_learn==1.3.0
scipy==1.10.1
seaborn>=0.11.0
statsmodels==0.14.0
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
gseapy
python-igraph
ndex2
tulip-python
matplotlib_venn
deepseek-ai>=0.1.0
python-dotenv>=0.19.0
